Defined method interface for new wizard architecture
In the Base.py file the new controller and view classes for a wizard entry have been defined.
The idea is that each stage in the wizard will consist of one WizardEntryController and its corresponding View.
This allows for more concise code as well as easier addition of new stages in the future.