Updating EditMeasurementWizard classes to use new architecture
The controller now has a list of WizardEntryControllers defined by a factory method.
The sub-controllers are responsible for updating and populating the gui.

In the main controller the way serialization works has now been changed as well.
The stages will store their contents in a dict, which the main controller will store on disk.
The views should not interact with disk anyway.