Prepared ToDo class to allow sweepable parameters
In the same file there is now also a `DictionaryWrapper` which functions as a pointer to a dictionary.
This is needed for all `ToDo`s in one sweep to be able to have a reference to the same dictionary, which will only be created once the first `ToDo` is run.