More intuitive z-axis selector
Instead of showing all possible parameters in the z-axis selector, it makes more sense to only show the ones unequal across the selected measurements.
This ensures that the plot actually shows something, where before it could happen, that a parameter equal across all measurements would be selected, and a blank plot would be generated.
It felt a bit 'buggy'...