Quality of Life features (#219)
* Selectable interpolation type in contour plot

If some of the selected measurements have more densely populated x-values than others, we need to pad the missing values somehow.
For now we simply added NAN values, but now we give the user the choice to linearly interpolate between the closest two existing values instead.

* Changeable plot axis bounds

This can be useful if measurements were performed in a bigger area than we are interested in.

* Colormaps for plots

The default colors for matplotlib are kind of ugly. Let\'s allow the user to choose their own colormap.

* Seperate Settings into panels

This commit also reorders the code and makes it more readable in PlotView

* Remove old comments

* Remove unused data from calculations when custom bounds are set

* Fix bug where empty z-data leads to exception

By checking if the measurement actually contains enough data along the z-axis we can stop an index error from happening

* Allow clamping of z-values

Some measurements generate noise which leads to very ugly artifacts when plotting. This commit allow the user to specify a range to which the data should be clamped.

* Remove z-axis bounds for line plot

* Add option to change plot title